+++ client/src/index.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import {Header} from './header.jsx';
-import {AnimalLocations} from './animalLocations.jsx';
+import {AnimalLocations} from './components/animalLocations/animalLocations.jsx';
 import {NewSighting} from './addSighting.jsx';
 import {Login} from './components/login/login.jsx';
 import {SightingAsText} from './components/sightingAsText/sightingAsText.jsx';
